+++ ASM/c/get_items.h
@@ -4,8 +4,8 @@
 #include <stdbool.h>
 #include "z64.h"
 
-extern uint16_t CFG_ADULT_TRADE_SHUFFLE;
-extern uint16_t CFG_CHILD_TRADE_SHUFFLE;
+extern uint8_t CFG_ADULT_TRADE_SHUFFLE;
+extern uint8_t CFG_CHILD_TRADE_SHUFFLE;
 
 void item_overrides_init();
 void handle_pending_items();
